Removed get absolute URL from abstract user and moved it into basic user. Fixed some small issues inside the user API view, adding serializer class and query set. Fixed type in API URLs. Added an env.dev.api file and changed docker-compose to read from it instead.
