+++ core/.env.dev.api
@@ -1,0 +1,1 @@
+DJANGOMODE="API"
