+++ core/AbstractModels/AbstractBaseUser/models.py
@@ -1,9 +1,6 @@
 """Abstract models for an abstracted basic user account."""
-from decimal import InvalidOperation
 from django.contrib.auth.models import AbstractUser, UserManager
 from django.db.models import EmailField, DateTimeField, BigAutoField
-from django.urls import reverse
-from django.conf import settings
 
 
 class AbstractBasicUser(AbstractUser):
@@ -21,7 +18,6 @@ class AbstractBasicUser(AbstractUser):
 
     Methods:
         __str__(self): Return name or username as a string
-        get_absolute_url(self): Returns the users edit url
     """
 
     objects = UserManager()
@@ -45,16 +41,3 @@ class AbstractBasicUser(AbstractUser):
         if self.get_full_name() == '':
             return self.username
         return f'{self.first_name} {self.last_name}'
-
-    @property
-    def get_absolute_url(self):
-        """Return the edit profile url."""
-        try:
-            if settings.RESPONSE_MODE == 'API':
-                return reverse('api_edit_account')
-            raise InvalidOperation('Invalid Response Mode')
-
-        except InvalidOperation:
-            print('\033[91m ERROR: INVALID RESPONSE_MODE SET IN YOUR SETTING.PY\033[0m')  # noqa: E501
-        except NameError:
-            print('\033[91m ERROR: RESPONSE_MODE NOT SET IN YOUR SETTING.PY\033[0m')  # noqa: E501
