+++ core/BasicUser/models.py
@@ -1,5 +1,9 @@
 """Model describing a Basic User for Django REST and templates."""
 from django.db import models
+from django.urls import reverse
+from django.conf import settings
+
+from decimal import InvalidOperation
 
 from AbstractModels.AbstractBaseUser.models import AbstractBasicUser
 
@@ -8,7 +12,20 @@ class BasicUser(AbstractBasicUser, models.Model):
     """Basic User is a bare bones basic user.
 
     Basic User wraps Abstract Base User allowing testing as well as views.
-    Class is passed.
+
+    Methods:
+        get_absolute_url: gets the url for api view user
     """
 
-    pass
+    @property
+    def get_absolute_url(self):
+        """Return the edit profile url."""
+        try:
+            if settings.RESPONSE_MODE == 'API':
+                return reverse('api_view_user', kwargs={'pk': self.id})
+            raise InvalidOperation('Invalid Response Mode')
+
+        except InvalidOperation:
+            print('\033[91m ERROR: INVALID RESPONSE_MODE SET IN YOUR SETTING.PY\033[0m')  # noqa: E501
+        except NameError:
+            print('\033[91m ERROR: RESPONSE_MODE NOT SET IN YOUR SETTING.PY\033[0m')  # noqa: E501
