+++ core/BasicUser/urls.py
@@ -12,6 +12,7 @@ router = DefaultRouter()
 urlpatterns = [
     path('auth/', include('knox.urls')),
     path('create/', UserAPI.as_view(), name='api_create_user'),
+    path('view/<pk>', UserAPI.as_view(), name='api_view_user'),
     path('login/', LoginAPI, name='api_login'),
     path('logout/', LogoutView.as_view(), name='api_logout_API'),
     path('edit/', UserAPI.as_view(), name='api_edit_account'),
