+++ core/docker-compose.yml
@@ -2,8 +2,6 @@ version: "3.9"
 
 services:
   api:
-    environment:
-        - DJANGOMODE="API"
     build: ./
     restart: always
     command: python manage.py runserver 0.0.0.0:8000
@@ -18,3 +16,5 @@ services:
         interval: 30s
         timeout: 10s
         retries: 5
+    env_file:
+      - ./.env.dev.api
